refactor
2021. 03. 10
ResultView 내부에 레이싱 진행 로직이 있어 잘못된 역할이라 생각되어, 해당 로직을 GrandPrix 로 옮겼습니다.
ResultView 는 순수 출력용 메소드만 남겼습니다.

Resolves : loop-study
See also : changgunyee
